LayoutTests/fast/dom/Geolocation/permission-denied-already-success-expected.txt

Tests that when Geolocation permission has been denied prior to a call to a Geolocation method, the error callback is invoked with code PERMISSION_DENIED, when the Geolocation service has a good position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"

PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS successfullyParsed is true

TEST COMPLETE
